requirements.txt
certifi==2017.7.27.1
emoji==0.4.5
feedparser==5.2.1
future==0.16.0
python-telegram-bot==8.1.1
six==1.11.0
